feat(YSP-000): rename link treatment
