+++ .storybook/preview.js
@@ -14,7 +14,7 @@ import '../fonts/fontawesome/css/fontawesome.css';
 import '../fonts/fontawesome/css/regular.css';
 import '../fonts/fontawesome/css/solid.css';
 
-import './link_treatment.js';
+import './link-treatment.js';
 
 // If in a Drupal project, it's recommended to import a symlinked version of drupal.js.
 import './_drupal.js';
